move run all games to scripts
