+++ brain_games/games/brain_gcd.py
@@ -1,15 +1,16 @@
 """Greatest Common Devizor game."""
 
 from brain_games.common import generate_random_int
-from brain_games.game_engine import run_game
 
 
-def brain_gcd() -> None:
-    """Implementaion of game's logic."""
-    run_game(
-        main_question='Find the greatest common divisor of given numbers.',
-        build_question=build_question_and_answer,
-    )
+def get_welcome_game_question() -> str:
+    """
+    Get welcome main game question to ask user at start.
+
+    Returns:
+        str
+    """
+    return 'Find the greatest common divisor of given numbers.'
 
 
 def build_question_and_answer() -> tuple:
