+++ brain_games/games/brain_progression.py
@@ -2,20 +2,20 @@
 
 from random import randint
 
-from brain_games.game_engine import run_game
-
 PROGRESSION_LEN_MIN = 5
 PROGRESSION_LEN_MAX = 10
 PROGRESSION_START_ITEM_MAX = 9
 PROGRESSION_ITERATOR_MAX = 5
 
 
-def brain_progression() -> None:
-    """Implementaion of game's logic."""
-    run_game(
-        main_question='What number is missing in the progression?',
-        build_question=build_question_and_answer,
-    )
+def get_welcome_game_question() -> str:
+    """
+    Get welcome main game question to ask user at start.
+
+    Returns:
+        str
+    """
+    return 'What number is missing in the progression?'
 
 
 def build_question_and_answer() -> tuple:
