+++ brain_games/scripts/brain_calc.py
@@ -1,12 +1,16 @@
 #!/usr/bin/env python
 """Game 'Calc'  entrypoint."""
 
-from brain_games.games.brain_calc import brain_calc
+from brain_games.game_engine import run_game
+from brain_games.games import brain_calc as game
 
 
 def main() -> None:
-    """Entrypoint function."""
-    brain_calc()
+    """Run game."""
+    run_game(
+        main_question=game.get_welcome_game_question(),
+        build_question=game.build_question_and_answer,
+    )
 
 
 if __name__ == '__main__':
