+++ brain_games/scripts/brain_even.py
@@ -1,12 +1,16 @@
 #!/usr/bin/env python
 """Game 'Even'  entrypoint."""
 
-from brain_games.games.brain_even import brain_even
+from brain_games.game_engine import run_game
+from brain_games.games import brain_even as game
 
 
 def main() -> None:
-    """Entrypoint function."""
-    brain_even()
+    """Run game."""
+    run_game(
+        main_question=game.get_welcome_game_question(),
+        build_question=game.build_question_and_answer,
+    )
 
 
 if __name__ == '__main__':
